fix(rpc_bench): make sure tcp stream being configured properly
We have moved rpc client from directly being used to wrap with
connection pools, and moved the tcp socket configuration out to be
called for different connection modes. However, the bench rpc is still
using the old style to create rpc clients, missing the opportunity to be
configured properly.

Note we may need to simplify and to clean up all the connection mode,
to avoid such bugs happen in the future.